Added font for icon of weather

+++ app/src/main/java/com/example/weatherapplication/modelWeather/Main.java
@@ -2,7 +2,7 @@ package com.example.weatherapplication.modelWeather;
 
 public class Main {
     private float temp;
-    private int pressure;
+    private float pressure;
     private int humidity;
     private float temp_min;
     private float temp_max;
@@ -15,7 +15,7 @@ public class Main {
         this.temp = temp;
     }
 
-    public int getPressure() {
+    public float getPressure() {
         return pressure;
     }
